+++ persistent/ChangeLog.md
@@ -1,3 +1,8 @@
+## 2.7.1
+
+* Added an `insertUniqueEntity` function [#718](https://github.com/yesodweb/persistent/pull/718)
+* Added `BackendCompatible` class [#701](https://github.com/yesodweb/persistent/pull/701)
+
 ## 2.7.0
 
 * Fix upsert behavior [#613](https://github.com/yesodweb/persistent/issues/613)